+++ plugins/woocommerce-admin/packages/components/src/select-control/index.js
@@ -15,36 +15,41 @@ import { withInstanceId, compose } from '@wordpress/compose';
  */
 import List from './list';
 import Tags from './tags';
-import SearchControl from './control';
+import Control from './control';
 
 /**
  * A search box which filters options while typing,
  * allowing a user to select from an option from a filtered list.
  */
-export class Autocomplete extends Component {
+export class SelectControl extends Component {
 	static getInitialState() {
 		return {
-			filteredOptions: [],
-			selectedIndex: 0,
+			isExpanded: false,
 			query: '',
 		};
 	}
 
 	constructor( props ) {
 		super( props );
-		this.state = this.constructor.getInitialState();
+		this.state = {
+			...this.constructor.getInitialState(),
+			filteredOptions: [],
+			selectedIndex: 0,
+		};
 
 		this.bindNode = this.bindNode.bind( this );
+		this.decrementSelectedIndex = this.decrementSelectedIndex.bind( this );
+		this.incrementSelectedIndex = this.incrementSelectedIndex.bind( this );
 		this.search = this.search.bind( this );
 		this.selectOption = this.selectOption.bind( this );
-		this.updateSelectedIndex = this.updateSelectedIndex.bind( this );
+		this.setExpanded = this.setExpanded.bind( this );
 	}
 
 	bindNode( node ) {
 		this.node = node;
 	}
 
-	reset( selected = this.props.selected ) {
+	reset( selected = this.getSelected() ) {
 		const { multiple } = this.props;
 		const initialState = this.constructor.getInitialState();
 
@@ -60,12 +65,6 @@ export class Autocomplete extends Component {
 		this.reset();
 	}
 
-	isExpanded() {
-		const { filteredOptions, query } = this.state;
-
-		return filteredOptions.length > 0 || query;
-	}
-
 	hasTags() {
 		const { multiple, selected } = this.props;
 
@@ -76,22 +75,54 @@ export class Autocomplete extends Component {
 		return selected.some( item => Boolean( item.label ) );
 	}
 
+	getSelected() {
+		const { multiple, options, selected } = this.props;
+
+		// Return the passed value if an array is provided.
+		if ( multiple || Array.isArray( selected ) ) {
+			return selected;
+		}
+
+		const selectedOption = options.find( option => option.key === selected );
+		return selectedOption ? [ selectedOption ] : [];
+	}
+
 	selectOption( option ) {
 		const { multiple, onChange, selected } = this.props;
 		const { query } = this.state;
 		const newSelected = multiple ? [ ...selected, option ] : [ option ];
 
-		// Check if this is already selected
-		const isSelected = findIndex( selected, { key: option.key } );
-		if ( -1 === isSelected ) {
-			onChange( newSelected, query );
+		// Trigger a change if the selected value is different and pass back
+		// an array or string depending on the original value.
+		if ( Array.isArray( selected ) ) {
+			const isSelected = findIndex( selected, { key: option.key } );
+			if ( -1 === isSelected ) {
+				onChange( newSelected, query );
+			}
+		} else if ( selected !== option.key ) {
+			onChange( option.key, query );
 		}
 
 		this.reset( newSelected );
 	}
 
-	updateSelectedIndex( value ) {
-		this.setState( { selectedIndex: value } );
+	decrementSelectedIndex() {
+		const { selectedIndex } = this.state;
+		const options = this.getOptions();
+		const nextSelectedIndex =
+			null !== selectedIndex
+				? ( selectedIndex === 0 ? options.length : selectedIndex ) - 1
+				: options.length - 1;
+
+		this.setState( { selectedIndex: nextSelectedIndex } );
+	}
+
+	incrementSelectedIndex() {
+		const { selectedIndex } = this.state;
+		const options = this.getOptions();
+		const nextSelectedIndex = null !== selectedIndex ? ( selectedIndex + 1 ) % options.length : 0;
+
+		this.setState( { selectedIndex: nextSelectedIndex } );
 	}
 
 	announce( filteredOptions ) {
@@ -117,9 +148,21 @@ export class Autocomplete extends Component {
 		}
 	}
 
+	getOptions() {
+		const { isSearchable, options } = this.props;
+		const { filteredOptions } = this.state;
+		return isSearchable ? filteredOptions : options;
+	}
+
 	getFilteredOptions( query ) {
-		const { excludeSelectedOptions, getSearchExpression, maxResults, onFilter, options, selected } = this.props;
-		const selectedKeys = selected.map( option => option.key );
+		const {
+			excludeSelectedOptions,
+			getSearchExpression,
+			maxResults,
+			onFilter,
+			options,
+		} = this.props;
+		const selectedKeys = this.getSelected().map( option => option.key );
 		const filtered = [];
 
 		// Create a regular expression to filter the options.
@@ -155,42 +198,50 @@ export class Autocomplete extends Component {
 		return onFilter( filtered );
 	}
 
+	setExpanded( value ) {
+		this.setState( { isExpanded: value } );
+	}
+
 	search( query ) {
 		const { hideBeforeSearch, onSearch, options } = this.props;
 
 		onSearch( query );
 		// Get all options if `hideBeforeSearch` is enabled and query is not null.
-		const filteredOptions = null !== query && ! query.length && ! hideBeforeSearch
-			? options
-			: this.getFilteredOptions( query );
-		this.setState( { selectedIndex: 0, filteredOptions, query: query || '' }, () => this.announce( filteredOptions ) );
+		const filteredOptions =
+			null !== query && ! query.length && ! hideBeforeSearch
+				? options
+				: this.getFilteredOptions( query );
+		this.setState(
+			{
+				selectedIndex: 0,
+				filteredOptions,
+				isExpanded: Boolean( filteredOptions.length ),
+				query: query || '',
+			},
+			() => this.announce( filteredOptions )
+		);
 	}
 
 	render() {
-		const {
-			className,
-			inlineTags,
-			instanceId,
-			options,
-		} = this.props;
-		const { selectedIndex } = this.state;
+		const { className, inlineTags, instanceId, isSearchable, options } = this.props;
+		const { isExpanded, selectedIndex } = this.state;
 
-		const isExpanded = this.isExpanded();
 		const hasTags = this.hasTags();
 		const { key: selectedKey = '' } = options[ selectedIndex ] || {};
-		const listboxId = isExpanded ? `woocommerce-autocomplete__listbox-${ instanceId }` : null;
+		const listboxId = isExpanded ? `woocommerce-select-control__listbox-${ instanceId }` : null;
 		const activeId = isExpanded
-			? `woocommerce-autocomplete__option-${ instanceId }-${ selectedKey }`
+			? `woocommerce-select-control__option-${ instanceId }-${ selectedKey }`
 			: null;
 
 		return (
 			<div
-				className={ classnames( 'woocommerce-autocomplete', className, {
+				className={ classnames( 'woocommerce-select-control', className, {
 					'has-inline-tags': hasTags && inlineTags,
+					'is-searchable': isSearchable,
 				} ) }
 				ref={ this.bindNode }
 			>
-				<SearchControl
+				<Control
 					{ ...this.props }
 					{ ...this.state }
 					activeId={ activeId }
@@ -198,26 +249,33 @@ export class Autocomplete extends Component {
 					isExpanded={ isExpanded }
 					listboxId={ listboxId }
 					onSearch={ this.search }
+					selected={ this.getSelected() }
+					setExpanded={ this.setExpanded }
+					decrementSelectedIndex={ this.decrementSelectedIndex }
+					incrementSelectedIndex={ this.incrementSelectedIndex }
 				/>
-				{ ! inlineTags && hasTags && <Tags { ...this.props } /> }
-				{ isExpanded &&
+				{ ! inlineTags && hasTags && <Tags { ...this.props } selected={ this.getSelected() } /> }
+				{ isExpanded && (
 					<List
 						{ ...this.props }
 						{ ...this.state }
 						activeId={ activeId }
 						listboxId={ listboxId }
 						node={ this.node }
-						onChange={ this.updateSelectedIndex }
 						onSelect={ this.selectOption }
 						onSearch={ this.search }
+						options={ this.getOptions() }
+						decrementSelectedIndex={ this.decrementSelectedIndex }
+						incrementSelectedIndex={ this.incrementSelectedIndex }
+						setExpanded={ this.setExpanded }
 					/>
-				}
+				) }
 			</div>
 		);
 	}
 }
 
-Autocomplete.propTypes = {
+SelectControl.propTypes = {
 	/**
 	 * Class name applied to parent div.
 	 */
@@ -238,14 +296,15 @@ Autocomplete.propTypes = {
 	/**
 	 * Help text to be appended beneath the input.
 	 */
-	help: PropTypes.oneOfType( [
-		PropTypes.string,
-		PropTypes.node,
-	] ),
+	help: PropTypes.oneOfType( [ PropTypes.string, PropTypes.node ] ),
 	/**
 	 * Render tags inside input, otherwise render below input.
 	 */
 	inlineTags: PropTypes.bool,
+	/**
+	 * Allow the select options to be filtered by search input.
+	 */
+	isSearchable: PropTypes.bool,
 	/**
 	 * A label to use for the main input.
 	 */
@@ -265,10 +324,7 @@ Autocomplete.propTypes = {
 	options: PropTypes.arrayOf(
 		PropTypes.shape( {
 			isDisabled: PropTypes.bool,
-			key: PropTypes.oneOfType( [
-				PropTypes.number,
-				PropTypes.string,
-			] ).isRequired,
+			key: PropTypes.oneOfType( [ PropTypes.number, PropTypes.string ] ).isRequired,
 			keywords: PropTypes.arrayOf( PropTypes.string ),
 			label: PropTypes.string,
 			value: PropTypes.any,
@@ -279,19 +335,19 @@ Autocomplete.propTypes = {
 	 */
 	placeholder: PropTypes.string,
 	/**
-	 * An array of objects describing selected values. If the label of the selected
-	 * value is omitted, the Tag of that value will not be rendered inside the
-	 * search box.
+	 * An array of objects describing selected values or optionally a string for a single value.
+	 * If the label of the selected value is omitted, the Tag of that value will not
+	 * be rendered inside the search box.
 	 */
-	selected: PropTypes.arrayOf(
-		PropTypes.shape( {
-			key: PropTypes.oneOfType( [
-				PropTypes.number,
-				PropTypes.string,
-			] ).isRequired,
-			label: PropTypes.string,
-		} )
-	),
+	selected: PropTypes.oneOfType( [
+		PropTypes.string,
+		PropTypes.arrayOf(
+			PropTypes.shape( {
+				key: PropTypes.oneOfType( [ PropTypes.number, PropTypes.string ] ).isRequired,
+				label: PropTypes.string,
+			} )
+		),
+	] ),
 	/**
 	 * A limit for the number of results shown in the options menu.  Set to 0 for no limit.
 	 */
@@ -314,10 +370,11 @@ Autocomplete.propTypes = {
 	staticList: PropTypes.bool,
 };
 
-Autocomplete.defaultProps = {
+SelectControl.defaultProps = {
 	excludeSelectedOptions: true,
 	getSearchExpression: identity,
 	inlineTags: false,
+	isSearchable: false,
 	onChange: noop,
 	onFilter: identity,
 	onSearch: noop,
@@ -333,4 +390,4 @@ export default compose( [
 	withSpokenMessages,
 	withInstanceId,
 	withFocusOutside, // this MUST be the innermost HOC as it calls handleFocusOutside
-] )( Autocomplete );
+] )( SelectControl );